requirements.txt

alabaster==0.7.10
altgraph==0.15
anyjson==0.3.3
argcomplete==1.8.1
argh==0.26.2
asn1crypto==0.24.0
attrs==17.2.0
Automat==0.6.0
Babel==2.5.3
backdoor-factory==0.0.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
BBQSQL==1.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
BlindElephant==1.0
blinker==1.3
brotlipy==0.6.0
capstone==3.0.4
certifi==2017.11.5
chardet==3.0.4
## FIXME: could not find svn URL in dependency_links for this package:
chirp===daily-20170714
click==6.7
colorama==0.3.7
ConfigArgParse==0.11.0
configobj==5.0.6
configparser==3.5.0
constantly==15.1.0
construct==2.8.16
couchdbkit==0.6.5
cryptography==2.1.4
cssutils==1.0.2
decorator==4.1.2
dicttoxml==1.7.4
dis3==0.1.2
dissy==9
distorm3==3.3.4
dnslib==0.9.7
dnspython==1.15.0
docutils==0.14
easygui==0.96
enum34==1.1.6
fbchat==1.2.1
feedparser==5.2.1
Flask==0.12.2
fuse-python==0.2.1
future==0.15.2
futures==3.2.0
GeoIP==1.3.2
gevent==1.2.2
Ghost.py==2.0.0.dev0
greenlet==0.4.12
h2==2.5.2
hpack==3.0.0
html2text==2016.9.19
html5lib==0.999999999
http-parser==0.8.3
httplib2==0.9.2
hyperframe==4.0.1
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
imgkit==0.1.8
impacket==0.9.15
incremental==16.10.1
ipaddress==1.0.17
IPy==0.83
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jdcal==1.0
Jinja2==2.10
jsbeautifier==1.6.4
jsonpickle==0.9.5
jsonrpclib==0.1.7
keepnote==0.7.8
keyring==10.6.0
keyrings.alt==2.2
killerbee==1.0
lxml==4.1.0
M2Crypto==0.27.0
macholib==1.9
Mako==1.0.7
MarkupSafe==1.0
mechanize==0.2.5
mercurial==4.4.1
metaconfig==0.1.4a1
mitmproxy==0.18.2
mockito==0.5.2
msgpack-python==0.4.8
mysqlclient==1.3.10
nassl==0.12
netaddr==0.7.19
NfSpy==1.0
numpy==1.13.3
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.3.0
PAM==0.4.2
paramiko==2.0.0
passlib==1.7.1
pathlib2==2.3.0
pathtools==0.1.2
pcapy==0.10.8
pefile==2017.11.5
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
ply==3.9
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.7.3
py==1.4.34
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycairo==1.15.4
pycrypto==2.6.1
pycurl==7.43.0
pydns==2.3.6
pyenchant==2.0.0
Pygments==2.2.0
pygobject==3.26.1
pygtkspellcheck==4.0.5
pyinotify==0.9.6
PyInstaller==3.3
pylibemu==0.3.3
pymongo==3.5.1
pymssql==2.1.3
pyOpenSSL==17.5.0
pyparsing==2.1.10
PyPDF2==1.26.0
pyperclip==1.6.0
pyqt==0.1
pyrit==0.5.1
pyscard==1.9.6
pyserial==3.4
PySide==1.2.2
pysnmp==4.3.2
pysnmp-apps==0.3.2
pysnmp-mibs==0.1.3
PySocks==1.6.5
pysqlite==2.7.0
pytesseract==0.1.7
pytest==3.2.1
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.2
pyusb==1.0.0b2
PyX==0.12.1
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
requests==2.18.4
restkit==4.2.2
rfidiot==1.0
roman==2.0.0
scandir==1.6
scapy==2.3.3
scipy==1.0.0
scour==0.36
SecretStorage==2.3.1
selenium==3.8.1
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
slowaes==0.1a1
snowballstemmer==1.2.1
socketpool==0.5.3
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.11
tcpwatch==1.3.1
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
urwid==1.3.1
uTidylib==0.3
vinetto==0.7b0
virtualenv==15.1.0
volatility==2.6
wapiti==2.3.0
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5
webkit2png==0.8.2
webscreenshot==2.1
webunit==1.3.10
Werkzeug==0.14.1
wfuzz==2.2.3
Whoosh==2.7.4
wxPython==3.0.2.0
wxPython-common==3.0.2.0
XlsxWriter==0.9.6
xmlbuilder==1.0
xvfbwrapper==0.2.8
yara-python==3.7.0
zenmap==7.60
zim==0.67
zope.interface==4.3.2
